some refactor so goplugins use the same logic even if they are middleware or response handlers
